Pruning taxa that are ancestors only of flag-pruned
This is deals with a long-standing issue that is much
more severe now that OTT 2.10 hides IRMNG-only taxa.
If all of the children of a genus are hidden, the genus
is not flagged as hidden by smasher. So, pruning based on
flags turns some (now many) higher level taxa into empty tips.

This commit uses new logic in otcetera and peyotl branches
	called prune-higher-tips right now) to support
a second pruning of OTT after the flag-based pruning.
Currently otcetera only prunes tips mapped to genus, species group,
or subgenus. It would be easy to make that more flexible.